+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Enhanced ConstructEditor:
 - Implemented "Copy" / Ctrl+C (#18)
 - Added "Copy path to clipboard" button in the context menu (#18)
 - Fixed a bug, when a struct has multiple times the same `Enum` construct. Then the metadata (eg. byte position) of the last parsed enum value is used for all enum values.
+- Added exception dialog to show the complete parse/build exception.
 
 Enhanced HexEditor:
 - fix crash when selecting or extending selection before the beginning of the hex editor using the shift LEFT and UP arrow keys (#20)